web_app/requirements.txt
# Core dependencies
numpy>=1.23.5,<2
opencv-python==4.11.0.86
onnx==1.17.0
insightface==0.7.3
pillow==11.1.0
torch; sys_platform != 'darwin' --index-url https://download.pytorch.org/whl/cu126
torch; sys_platform == 'darwin' --index-url https://download.pytorch.org/whl/cu126
torchvision; sys_platform != 'darwin' --index-url https://download.pytorch.org/whl/cu126
torchvision; sys_platform == 'darwin' --index-url https://download.pytorch.org/whl/cu126
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.21; sys_platform != 'darwin'
tensorflow; sys_platform != 'darwin'
gfpgan==1.3.8

# Web version dependencies
flask==2.0.1
werkzeug==2.0.3
flask-socketio==5.1.1
python-engineio==4.2.1
python-socketio==5.4.0
eventlet==0.33.0
dnspython==2.2.1 